Bug fixes and removed conflicting files.
